refactor: change package infrastructure
* clean-up reporting
* only show specific package version in report meta information
