+++ MANIFEST.in
@@ -1,12 +1,6 @@
 include AUTHORS.rst
-include CONTRIBUTING.rst
 include HISTORY.rst
 include LICENSE
 include README.rst
 
 recursive-include memote *.py *.html *.yml
-recursive-include tests *.py
-recursive-exclude * __pycache__
-recursive-exclude * *.py[co]
-
-recursive-include docs *.rst conf.py Makefile make.bat *.jpg *.png *.gif
