--- memote/suite/reporting/templates/basic_report.html
@@ -1,37 +1,0 @@
-{% extends "layout.html" %}
-{% from '_card_helpers.html' import render_card_content %}
-
-{% block header -%}
-    <div class="md-toolbar-tools">
-        <h1 class="md-toolbar-tools" flex>{{ super() }} - {{ timestamp }}</h1>
-    </div>
-{%- endblock %}
-
-{% block content -%}
-    <md-content layout="row" class="md-padding" layout-xs="column">
-        <div layout="column" flex>
-            {{ render_card_content(
-                'Basic Information',
-                data['report']['test_basic']
-            ) }}
-            {{ render_card_content(
-                'Stoichiometric Consistency',
-                data['report']['test_consistency']
-            ) }}
-            {{ render_card_content(
-                'Syntax Information',
-                data['report']['test_syntax']
-            ) }}
-        </div>
-        <div layout="column" flex>
-            {{ render_card_content(
-                'Biomass Production',
-                data['report']['test_biomass']
-            ) }}
-            {{ render_card_content(
-                'Annotation Information',
-                data['report']['test_annotation']
-            ) }}
-        </div>
-    </md-content>
-{%- endblock %}
