--- memote/suite/reporting/templates/history_report.html
@@ -1,57 +1,0 @@
-{% extends "layout.html" %}
-{% from '_card_helpers.html' import render_plot_card %}
-
-{% block title -%}
-<title>{{ names|join(", ") }} Report</title>
-{%- endblock %}
-
-{% block topscript -%}
-    <script src="https://cdn.rawgit.com/mathjax/MathJax/2.7.1/MathJax.js?config=TeX-AMS-MML_HTMLorMML&delayStartupUntil=configured"></script>
-    <script src="https://cdn.plot.ly/plotly-1.26.1.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/d3/3.3.5/d3.js" charset="utf-8"></script>
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
-{%- endblock %}
-
-{% block header -%}
-{{ names|join(", ") }} Report
-{%- endblock %}
-
-{% block content -%}
-    <md-content layout="row" class="md-padding" layout-xs="column">
-        <div layout="column" flex>
-            {{ render_plot_card('Basic Information', basics) }}
-
-            {{ render_plot_card('Stoichiometric Consistency', consistency) }}
-
-{#             {{ render_plot_card('Syntax Information', basics) }} #}
-            <md-card ng-app>
-                <md-card-title>
-                    <span class="md-headline">Syntax Information</span>
-                </md-card-title>
-            </md-card>
-        </div>
-        <div layout="column" flex>
-            {{ render_plot_card('Biomass Production', biomass) }}
-
-{#           {{ render_plot_card('Annotation Information', basics) }} #}
-            <md-card ng-app>
-                <md-card-title>
-                    <span class="md-headline">Annotation Information</span>
-                </md-card-title>
-            </md-card>
-        </div>
-    </md-content>
-{%- endblock %}
-
-
-{% block bottomscript %}
-{{ super() }}
-<script>
-    var waitForPlotly = setInterval( function() {
-        if( typeof(window.Plotly) !== "undefined" ){
-            MathJax.Hub.Config({ SVG: { font: "STIX-Web" }, displayAlign: "center" });
-            MathJax.Hub.Queue(["setRenderer", MathJax.Hub, "SVG"]);
-            clearInterval(waitForPlotly);
-        }}, 250 );
-</script>
-{%- endblock %}
