+++ pyproject.toml
@@ -1,0 +1,5 @@
+[build-system]
+requires = [
+  "setuptools>=30.3.0",
+  "wheel"
+]
