+++ setup.cfg
@@ -1,3 +1,40 @@
+[metadata]
+name = memote
+version = 0.5.1
+url = https://github.com/opencobra/memote
+download_url = https://pypi.python.org/pypi/memote/
+author = Moritz E. Beber
+author_email = morbeb@biosustain.dtu.dk
+classifiers =
+    Development Status :: 5 - Production/Stable
+    Intended Audience :: Developers
+    Intended Audience :: Science/Research
+    Topic :: Scientific/Engineering :: Bio-Informatics
+    License :: OSI Approved :: Apache Software License
+    Natural Language :: English
+    Programming Language :: Python :: 2
+    Programming Language :: Python :: 2.7
+    Programming Language :: Python :: 3
+    Programming Language :: Python :: 3.5
+    Programming Language :: Python :: 3.6
+license = Apache Software License Version 2.0
+description = the genome-scale metabolic model test suite
+long_description = file: README.rst
+keywords =
+    memote
+    metabolic
+    constained-based
+    model
+    COBRA
+    test suite
+    quality control
+    quality assurance
+
+[options]
+packages = find:
+include_package_data = True
+zip_safe = False
+
 [bumpversion]
 current_version = 0.5.1
 commit = True
@@ -18,9 +55,9 @@ values =
 	placeholder
 	a
 
-[bumpversion:file:setup.py]
-search = version="{current_version}"
-replace = version="{new_version}"
+[bumpversion:file:setup.cfg]
+search = version = {current_version}
+replace = version = {new_version}
 
 [bumpversion:file:memote/__init__.py]
 search = __version__ = "{current_version}"
@@ -31,6 +68,9 @@ search = "{current_version}"
 replace = "{new_version}"
 
 [wheel]
+universal = 1
+
+[bdist_wheel]
 universal = 1
 
 [flake8]
