+++ test_requirements.txt
@@ -1,0 +1,4 @@
+pytest>=3.1
+pytest-raises
+pytest-cov
+codecov
