+++ tox.ini
@@ -1,13 +1,10 @@
 [tox]
-envlist = flake8, docs, py27, py34, py35, py36
+envlist = flake8, docs, py{27,35,36}
 
 [testenv]
 usedevelop = True
 deps =
-    pytest>=3.1
-    pytest-raises
-    pytest-cov
-    codecov
+    -r{toxinidir}/test_requirements.txt
 passenv =
     CI
     TRAVIS
